requirements.txt
# NEAT-Pong-Python Requirements
# Enhanced Version - December 2025

# Core dependencies
neat-python>=0.92
pygame>=2.0.0

# Optional for data analysis
# pandas>=1.3.0
# matplotlib>=3.3.0

# Development tools (optional)
# pytest>=6.2.0
# pylint>=2.8.0
